# core/vm: 64 bit gas instructions