This PR changes all gas instructions from `*big.Int` to `uint64`, a new theoretical limit.

It also implements a new `*big.Int` instruction pool, which contains reusable `*big.Int`s. Instructions found in `core/vm/instructions.go` do no longer initialise new `big.Int`s but instead call `intPool.Get()` returning either a reusable integer -- **beware** this may contain an old value -- or `new(big.Int)` if the pool is empty. I've decided to implement my own pool instead of using the `sync.Pool` because I do not require a thread safe pool and I do not want the typecasting overhead required with the `sync.Pool.