@obscuren 2017-01-13
```
benchmark                         old ns/op     new ns/op     delta
BenchmarkVmAckermann32Tests-4     3214842       2171613       -32.45%
BenchmarkVmFibonacci16Tests-4     19069610      13485337      -29.28%

benchmark                         old allocs     new allocs     delta
BenchmarkVmAckermann32Tests-4     39283          34590          -11.95%
BenchmarkVmFibonacci16Tests-4     228242         203283         -10.94%

benchmark                         old bytes     new bytes     delta
BenchmarkVmAckermann32Tests-4     1478684       1340153       -9.37%
BenchmarkVmFibonacci16Tests-4     8444250       7935278       -6.03%