@holiman 2017-01-23
Regarding the intpool... It looks a bit 'dangerous', in that it seems relatively easy to make a mistake there. I feel it would be good to have some additional tests for coverage; not only to ensure that no consensus error occurs, but specifically it would be good to run the VM in 'checked' mode where it performs additional tests on the pool. 

Examples of what could be done in checked mode:
 
- Aggressively set a re-pooled bigint to e.g `-42`. Should make it more prone to fail in case of misuse. 
- Aggressively validate that all pooled bigints are `-42`, to detect outside modification (e.g. after each pc step or after each transaction).

I'm also concerned whether journal reverts can somehow hold a reference to a pooled bigint. And how we can ensure that such case does not occur in the future (because from what I've seen now, all journal entries appear to instantiate new bigints). 

Third, it probably makes sense to have some rudimentary counter on the pool - it does not appear limited? If you run a node for several weeks, and there's a skewed balance (if there are more pushing than popping), couldn't that eventually become a problem.. ? The `PUSH` instruction, for example, creates new bigints (https://github.com/obscuren/go-ethereum/blob/gas64/core/vm/instructions.go#L709)  instead of reusing existing ones.