@obscuren 2017-01-23
> I feel it would be good to have some additional tests for coverage

I agree, I'll work on those.

> I'm also concerned whether journal reverts can somehow hold a reference to a pooled bigint

If that were the case we'd already have issues (*). I'm going to assume it's not the case and that our current network would have already had ran in to any issues.

> Third, it probably makes sense to have some rudimentary counter on the pool - it does not appear limited?

(*) The int pool is being recreated each time a interpreter is instantiated, which happens for each transaction. I'm not sure how the pool can grow very large either. Ints must come from either a `PUSH` instruction or the auxiliary instructions.