@obscuren 2017-01-25
EVM integer pool integrity verifier has been added @holiman: `go build ./cmd/geth -tags VERIFY_EVM_INTEGER_POOL`. Bear in mind, it's very slow as each and every integer in the pool needs to be verified (big int comparison) for each instruction.