@obscuren 2017-02-01
Before anyone merges, ping me. I want to clean up the commit message.